requirements.txt
# Web framework and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0

# Templating
jinja2>=3.1.0

# Form and multipart handling
python-multipart>=0.0.6

# Session middleware (Starlette sessions)
itsdangerous>=2.1.0

# Password hashing (auth)
bcrypt>=4.0.0

# Environment variables (.env)
python-dotenv>=1.0.0

# Testing (Stage 10)
pytest>=7.4.0
httpx>=0.27.0
